config: run :platform:ui's and :platform:ui-test's tests on the emulator
