+++ platform/ui-test/src/androidTest/AndroidManifest.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest xmlns:android="http://schemas.android.com/apk/res/android">
+    <application>
+        <activity
+            android:name="com.jeanbarrossilva.orca.platform.ui.test.core.SingleFragmentActivityTests$NoDestinationActivity" />
+        <activity
+            android:name="com.jeanbarrossilva.orca.platform.ui.test.core.SingleFragmentActivityTests$InequivalentlyRoutedDestinationActivity" />
+        <activity
+            android:name="com.jeanbarrossilva.orca.platform.ui.test.core.SingleFragmentActivityTests$NonFragmentDestinationActivity" />
+        <activity
+            android:name="com.jeanbarrossilva.orca.platform.ui.test.core.SingleFragmentActivityTests$MultipleDestinationsActivity" />
+        <activity
+            android:name="com.jeanbarrossilva.orca.platform.ui.test.core.SingleFragmentActivityTests$PosteriorlyAddedDestinationActivity" />
+    </application>
+</manifest>
