+++ platform/ui-test/src/androidTest/java/com/jeanbarrossilva/orca/platform/ui/test/core/test/TestScope.extensions.kt
@@ -1,0 +1,22 @@
+package com.jeanbarrossilva.orca.platform.ui.test.core.test
+
+import kotlin.coroutines.resume
+import kotlin.coroutines.suspendCoroutine
+import kotlinx.coroutines.test.runTest
+
+/**
+ * Waits until the result of [condition] is `true`.
+ *
+ * @param condition Returns the expectation to be met.
+ **/
+internal fun waitUntil(condition: () -> Boolean) {
+    runTest {
+        suspendCoroutine {
+            @Suppress("ControlFlowWithEmptyBody")
+            while (!condition()) {
+            }
+
+            it.resume(Unit)
+        }
+    }
+}
