+++ platform/ui-test/src/main/java/com/jeanbarrossilva/orca/platform/ui/test/core/NavController.extensions.kt
@@ -1,0 +1,12 @@
+package com.jeanbarrossilva.orca.platform.ui.test.core
+
+import androidx.navigation.NavController
+
+/** Whether this [NavController] has a [graph][NavController.graph]. **/
+internal val NavController.hasNavGraph
+    get() = try {
+        graph
+        true
+    } catch (_: IllegalStateException) {
+        false
+    }
