+++ platform/ui/src/androidTest/java/com/jeanbarrossilva/orca/platform/ui/component/stat/TestActivateableStatIcon.kt
