+++ platform/ui/src/androidTest/java/com/jeanbarrossilva/orca/platform/ui/component/timeline/toot/test/SemanticsNodeInteraction.extensions.kt
@@ -1,0 +1,9 @@
+package com.jeanbarrossilva.orca.platform.ui.component.timeline.toot.test
+
+import androidx.compose.ui.test.SemanticsNodeInteraction
+import androidx.compose.ui.test.onChildAt
+
+/** [SemanticsNodeInteraction] of this [Stat]'s label node. **/
+internal fun SemanticsNodeInteraction.onStatLabel(): SemanticsNodeInteraction {
+    return onChildAt(1)
+}
