+++ platform/ui/src/androidTest/java/com/jeanbarrossilva/orca/platform/ui/core/lifecycle/test/Activity.extensions.kt
@@ -1,0 +1,64 @@
+package com.jeanbarrossilva.orca.platform.ui.core.lifecycle.test
+
+import android.app.Activity
+import android.app.Application
+import android.os.Bundle
+import kotlin.reflect.full.declaredMemberFunctions
+import kotlin.reflect.jvm.isAccessible
+
+/**
+ * Runs the [action] when this [Activity] is destroyed.
+ *
+ * @param action Operation to be performed when this [Activity] is destroyed.
+ **/
+internal fun <T : Activity> T.doOnDestroy(action: T.() -> Unit) {
+    registerActivityLifecycleCallbacks(
+        object : Application.ActivityLifecycleCallbacks {
+            override fun onActivityCreated(activity: Activity, savedInstanceState: Bundle?) {
+            }
+
+            override fun onActivityStarted(activity: Activity) {
+            }
+
+            override fun onActivityResumed(activity: Activity) {
+            }
+
+            override fun onActivityPaused(activity: Activity) {
+            }
+
+            override fun onActivityStopped(activity: Activity) {
+            }
+
+            override fun onActivitySaveInstanceState(activity: Activity, outState: Bundle) {
+            }
+
+            override fun onActivityDestroyed(activity: Activity) {
+                @Suppress("UNCHECKED_CAST")
+                (activity as T).action()
+
+                unregisterActivityLifecycleCallbacks(this)
+            }
+        }
+    )
+}
+
+/**
+ * Calls this [Activity]'s [onStart][Activity.onStart] method.
+ **/
+internal inline fun <reified T : Activity> T.onStart() {
+    call("onStart")
+}
+
+/**
+ * Calls this [Activity]'s method named [methodName].
+ *
+ * @param methodName Name of the method to be called.
+ * @param params Parameters with which the method will be called.
+ **/
+private inline fun <reified T : Activity> T.call(methodName: String, vararg params: Any?) {
+    T::class
+        .declaredMemberFunctions
+        .first { it.name == methodName }
+        .apply { isAccessible = true }
+        .call(this, *params)
+}
